Create server_combined.c
This is the combined server.c file which has the logon, logoff and getFileFromDB functions commented out as these have been moved to museumDB.c we should still use the server.c for testing purposes and when using static data which is set in the server.c file but when dynamic data is to be used, i.e. pulling data from the database via the functions in museumDB.c then we should migrate to server_combined.c and server.c should be deprecated and made obsolete.