Create museuminfodb.sql
This is the MySQL Database creation script to create the DB on any MySQLServer. 
Basically it creates:
The database with a name of museuminfodb
Populates basic user data into the tables.
Created the museum_db_admin account for access to the database. I have left out the password in this script for security reasons